Functional abstraction for camera1
