+++ letsrobotandroid-api/src/main/java/tv/letsrobot/android/api/Core.kt
@@ -15,10 +15,7 @@ import okhttp3.Request
 import org.json.JSONException
 import org.json.JSONObject
 import tv.letsrobot.android.api.EventManager.Companion.TIMEOUT
-import tv.letsrobot.android.api.components.AudioComponent
-import tv.letsrobot.android.api.components.CommunicationComponent
-import tv.letsrobot.android.api.components.RobotControllerComponent
-import tv.letsrobot.android.api.components.TextToSpeechComponent
+import tv.letsrobot.android.api.components.*
 import tv.letsrobot.android.api.components.api19.CameraComponent
 import tv.letsrobot.android.api.enums.CommunicationType
 import tv.letsrobot.android.api.enums.ComponentStatus
@@ -46,7 +43,7 @@ class Core
 private constructor(val robotId : String, val cameraId : String?) {
     private val handlerThread: HandlerThread = HandlerThread("bg-thread")
     private var logLevel = LogLevel.NONE
-    var camera: CameraComponent? = null
+    var camera: CameraBaseComponent? = null
     var audio: AudioComponent? = null
     var robotController: RobotControllerComponent? = null
     var textToSpeech: TextToSpeechComponent? = null
