+++ letsrobotandroid-api/src/main/java/tv/letsrobot/android/api/components/api19/CameraComponent.kt
@@ -3,23 +3,12 @@ package tv.letsrobot.android.api.components.api19
 import android.content.Context
 import android.graphics.ImageFormat
 import android.graphics.Rect
-import android.graphics.YuvImage
 import android.hardware.Camera
 import android.util.Log
 import android.view.SurfaceHolder
-import com.github.hiteshsondhi88.libffmpeg.FFmpeg
 import com.github.hiteshsondhi88.libffmpeg.FFmpegExecuteResponseHandler
-import com.github.hiteshsondhi88.libffmpeg.exceptions.FFmpegCommandAlreadyRunningException
-import com.google.common.util.concurrent.RateLimiter
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import org.json.JSONException
-import org.json.JSONObject
-import tv.letsrobot.android.api.enums.ComponentStatus
-import tv.letsrobot.android.api.interfaces.Component
-import tv.letsrobot.android.api.utils.StoreUtil
+import tv.letsrobot.android.api.components.CameraBaseComponent
 import java.io.IOException
-import java.util.concurrent.atomic.AtomicBoolean
 
 
 /**
@@ -37,89 +26,12 @@ class CameraComponent
  * @param context Needed to access the camera
  * @param cameraId camera id for robot
  */
-constructor(context: Context, val cameraId: String, val holder: SurfaceHolder) : Component(context), FFmpegExecuteResponseHandler, android.hardware.Camera.PreviewCallback, SurfaceHolder.Callback {
-    internal var ffmpegRunning = AtomicBoolean(false)
-    var ffmpeg : FFmpeg
+constructor(context: Context, cameraId: String, val holder: SurfaceHolder) : CameraBaseComponent(context, cameraId), FFmpegExecuteResponseHandler, android.hardware.Camera.PreviewCallback, SurfaceHolder.Callback {
+    private lateinit var r: Rect
+
     init {
         holder.addCallback(this)
-        ffmpeg = FFmpeg.getInstance(context)
     }
-    var UUID = java.util.UUID.randomUUID().toString()
-    var process : Process? = null
-    var port: String? = null
-    var host: String? = null
-    var streaming = AtomicBoolean(false)
-    var previewRunning = false
-    override fun enableInternal(){
-        try {
-            val client = OkHttpClient.Builder()
-                    .build()
-            var call = client.newCall(Request.Builder().url(String.format("https://letsrobot.tv/get_video_port/%s", cameraId)).build())
-            var response = call.execute()
-            if (response.body() != null) {
-                val `object` = JSONObject(response.body()!!.string())
-                Log.d("ROBOT", `object`.toString())
-                port = `object`.getString("mpeg_stream_port")
-            }
-            call = client.newCall(Request.Builder().url(String.format("https://letsrobot.tv/get_websocket_relay_host/%s", cameraId)).build())
-            response = call.execute()
-            if (response.body() != null) {
-                val `object` = JSONObject(response.body()!!.string())
-                Log.d("ROBOT", `object`.toString())
-                host = `object`.getString("host")
-            }
-        } catch (e: IOException) {
-            e.printStackTrace()
-        } catch (e: JSONException) {
-            e.printStackTrace()
-        }
-        if(host == null || port == null){
-            status = ComponentStatus.ERROR
-        }
-        else
-            streaming.set(true)
-    }
-
-    private var successCounter: Int = 0
-
-    fun bootFFMPEG(){
-        if(!streaming.get()){
-            ffmpegRunning.set(false)
-            status = ComponentStatus.DISABLED
-            return
-        }
-        successCounter = 0
-        status = ComponentStatus.CONNECTING
-        try {
-            // to execute "ffmpeg -version" command you just need to pass "-version"
-            val xres = "640"
-            val yres = "480"
-
-            val rotationOption = StoreUtil.getOrientation(context).ordinal //leave blank
-            val builder = StringBuilder()
-            for (i in 0..rotationOption){
-                if(i == 0) builder.append("-vf transpose=1")
-                else builder.append(",transpose=1")
-            }
-            print("\"$builder\"")
-            val kbps = "20"
-            val video_host = host
-            val video_port = port
-            val stream_key = StoreUtil.getCameraPass(context)
-            //TODO hook up with bitrate and resolution prefs
-            val command = "-f image2pipe -codec:v mjpeg -i - -f mpegts -framerate 30 -codec:v mpeg1video -b:v 10k -bf 0 -muxdelay 0.001 -tune zerolatency -preset ultrafast -pix_fmt yuv420p $builder http://$video_host:$video_port/$stream_key/$xres/$yres/"
-            ffmpeg.execute(UUID, null, command.split(" ").toTypedArray(), this)
-        } catch (e: FFmpegCommandAlreadyRunningException) {
-            e.printStackTrace()
-            // Handle if FFmpeg is already running
-        }
-    }
-
-    var width = 0
-    var height = 0
-    var limiter = RateLimiter.create(30.0)
-
-    private lateinit var r: Rect
 
     override fun onPreviewFrame(b: ByteArray?, camera: android.hardware.Camera?) {
         if(!streaming.get()) return
@@ -132,17 +44,7 @@ constructor(context: Context, val cameraId: String, val holder: SurfaceHolder) :
                 r = Rect(0, 0, width, height)
             }
         }
-        if (!ffmpegRunning.getAndSet(true)) {
-            bootFFMPEG()
-        }
-        process?.let {
-            val im = YuvImage(b, ImageFormat.NV21, width, height, null)
-            try {
-                im.compressToJpeg(r, 100, it.outputStream)
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
-        }
+        push(b, ImageFormat.NV21, r)
     }
 
     private fun setupCam(){
@@ -180,53 +82,6 @@ constructor(context: Context, val cameraId: String, val holder: SurfaceHolder) :
         streaming.set(false)
     }
 
-    override fun onStart() {
-        ffmpegRunning.set(true)
-        @Suppress("ConstantConditionIf")
-        if(shouldLog)
-            Log.d(LOGTAG, "onStart")
-    }
-
-    override fun onProgress(message: String?) {
-        @Suppress("ConstantConditionIf")
-        if(shouldLog)
-            Log.d(LOGTAG, "onProgress : $message")
-        successCounter++
-        status = when {
-            successCounter > 5 -> ComponentStatus.STABLE
-            successCounter > 2 -> ComponentStatus.INTERMITTENT
-            else -> ComponentStatus.CONNECTING
-        }
-    }
-
-    override fun onFailure(message: String?) {
-        Log.e(LOGTAG, "progress : $message")
-        status = ComponentStatus.ERROR
-    }
-
-    override fun onSuccess(message: String?) {
-        @Suppress("ConstantConditionIf")
-        if(shouldLog)
-            Log.d(LOGTAG, "onSuccess : $message")
-    }
-
-    override fun onFinish() {
-        @Suppress("ConstantConditionIf")
-        if(shouldLog)
-            Log.d(LOGTAG, "onFinish")
-        ffmpegRunning.set(false)
-        process?.destroy()
-        process = null
-        status = ComponentStatus.DISABLED
-    }
-
-    override fun onProcess(p0: Process?) {
-        @Suppress("ConstantConditionIf")
-        if(shouldLog)
-            Log.d(LOGTAG, "onProcess")
-        this.process = p0
-    }
-
     private var camera : android.hardware.Camera? = null
 
     private var surface = false
@@ -250,11 +105,5 @@ constructor(context: Context, val cameraId: String, val holder: SurfaceHolder) :
         camera?.setPreviewCallback (null)
         camera?.release()
         camera = null
-    }
-
-    companion object {
-        private const val LOGTAG = "CameraComponent"
-        private const val shouldLog = true
-        private val cameraActive = AtomicBoolean(false)
     }
 }
